eval-harness/pyproject.toml
[project]
name = "helpdesk-eval"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9",
    "anthropic>=0.49",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[tool.uv]
package = false
